requirements.txt
numpy==1.15.4
xmltodict==0.11.0
pdfrw==0.4
setuptools==40.5.0
decorator==4.3.0
Cython==0.29
reportlab==3.5.10
matplotlib==3.0.0
nose==1.3.7
biomart==0.9.2
CacheControl==0.12.5
ete3==3.1.1
Sphinx==1.8.1
requests==2.20.1
pandas==0.23.4
scipy==1.1.0
ipython==7.2.0
python_dateutil==2.7.5
hdmedians==0.13
httpretty==0.9.6
natsort==5.5.0
pytest==4.0.2
scikit_learn==0.20.1
sphinx_bootstrap_theme==0.6.5
